+++ .prettierrc.json
@@ -2,5 +2,13 @@
   "tabWidth": 2,
   "endOfLine": "auto",
   "arrowParens": "avoid",
-  "singleQuote": true
+  "singleQuote": true,
+  "overrides": [
+    {
+      "files": "*.scss",
+      "options": {
+        "tabWidth": 4
+      }
+    }
+  ]
 }
