+++ src/pages/member/Componenets/LoginForm.scss
@@ -1,0 +1,20 @@
+.loginForm {
+    .loginName .loginPwName {
+        font-size: 14px;
+        font-weight: 500;
+        line-height: 1em;
+        color: rgb(0, 0, 0);
+    }
+
+    input {
+        display: block;
+        width: 100%;
+        font-weight: 400;
+        letter-spacing: -0.15px;
+        border: 1px solid rgb(221, 224, 226);
+        background-color: rgb(255, 255, 255);
+        height: 48px;
+        box-sizing: border-box;
+        padding: 0px 16px;
+    }
+}
